requirements.txt
smbus==1.1.post2
paho-mqtt==1.6.1